requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
discord==1.7.3
discord.py==1.7.3
h11==0.13.0
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.23.0
outcome==1.2.0
pathspec==0.9.0
phpserialize==1.3
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.2
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-slugify==6.1.2
requests==2.28.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
undetected-chromedriver==3.1.5.post4
urllib3==1.26.9
websockets==10.3
wsproto==1.1.0
yarl==1.7.2
